RematesQA/pyproject.toml
[tool.poetry]
name = "rematesqa"
version = "0.1.0"
description = ""
authors = ["FelepeX"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyleniumio = "^1.20.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
